remote: sync messages in TRASH or SPAM as well (#35)
Messages are deleted after 30 days in GMail so if you add the trash (or
maybe spam) labels this will cause the messages to be deleted at some
point (this is probably already the current behavior). A full sync (`gmi
pull -f`) is required to get all spam'ed and trash'ed messages, consider
doing a dry-run first. Push support was already working so ideally this
should not really overwrite local changes.