ignore custom tags when pushing
Specify custom tags to be ignored when pushing:

  gmi set --ignore-tags queue,todo

Separate tags with comma. you have to specify the full list if you want
to modify. Use:

  gmi set

to see current set. You can clear the ignored tags with:

  gmi set --ignore-tags ""

the change will not work retrospectively, and you will have to manually
change something on a message for it to be pushed anew.