gmailieer: behave more like sendmail
sendmail(1) allows additional recipients to be passed in the command
line. The path to the message is never passed, it's always stdin:

> sendmail [option ...] [recipient ...]

At least neomutt makes use of the `recipient` arguments (and doesn't set
-t), so lieer needs to handle this properly.

Most sendmail implementations allow passing additional recipients in additional
arguments, but there's no way to express this to the GMail API - it basically
only supports the `-t` (`--read-recipients`) mode of sendmail, without
additional recipients.

However, MUA can make use of both `-t` and additional recipients, even in a
redundant fashion (with `-t` set, providing recipients in the message AND as
recipient in the CLI).

We need to check if the MUA asks us to do something we can't do.

The following combinations are OK:

 - When `-t` is passed, we need to check for the CLI-passed recipients to be
   equal or a subset of the ones passed in the headers.

 - When `-t` is not passed, all header-passed recipients need to be provided in
   the CLI as well.

This avoids silently not sending mail to some recipients (pretending we did),
or sending mail to recipients we didn't want to send to again.

Fixes #152