gmailieer: behave like sendmail
sendmail(1) allows additional recipients to be passed in the command
line. The path to the message is never passed, it's always stdin:

> sendmail [option ...] [recipient ...]

At least neomutt makes use of the `recipient` arguments, so let's have
lieer behave properly.

Contrary to SMTP, it seems recipients can only specified inside the
message body when interacting with GMail.
If additional recipients are passed, we add them to the Bcc field.

Fixes #152