Fix gmi send -d (dry-run)
msg within send function is used to read in stdin first. Later on in the
function msg is used to get back data from the message send, however
this occurs inside a `not dry-run` condition.

When a dry-run is made, msg is not updated and holds stale data from
stdin.read. The function then fails trying to print the stale msg data.

Fix by seperating into two print statements, one for the dry-run, and
one without.